+++ src/settings.ts
@@ -4,5 +4,6 @@ const rammerhead = true;
 const github = "https://github.com/cognetwork-dev/Metallic";
 const discord = "https://discord.com/invite/yk33HZSZkU";
 const privacyPolicyUpdated = 1707797140659;
+const gamesCdn = "https://radon.games";
 
-export { bare, obfuscate, rammerhead, github, discord, privacyPolicyUpdated };
+export { bare, obfuscate, rammerhead, github, discord, privacyPolicyUpdated, gamesCdn };